refactor botocore integration
